+++ GoogleAdManager/src/main/java/com/applovin/mediation/adapters/GoogleAdManagerMediationAdapter.java
@@ -1160,6 +1160,7 @@ public class GoogleAdManagerMediationAdapter
                     View mediaView = null;
                     MediaContent mediaContent = nativeAd.getMediaContent();
                     List<NativeAd.Image> images = nativeAd.getImages();
+                    Drawable mainImage = null;
                     float mediaContentAspectRatio = 0.0f;
 
                     if ( mediaContent != null )
@@ -1168,6 +1169,7 @@ public class GoogleAdManagerMediationAdapter
                         googleMediaView.setMediaContent( mediaContent );
                         mediaView = googleMediaView;
 
+                        mainImage = mediaContent.getMainImage();
                         mediaContentAspectRatio = mediaContent.getAspectRatio();
                     }
                     else if ( images != null && images.size() > 0 )
@@ -1207,6 +1209,11 @@ public class GoogleAdManagerMediationAdapter
                             .setBody( nativeAd.getBody() )
                             .setMediaView( mediaView )
                             .setCallToAction( nativeAd.getCallToAction() );
+
+                    if ( AppLovinSdk.VERSION_CODE >= 11_04_03_99 )
+                    {
+                        builder.setMainImage( new MaxNativeAd.MaxNativeAdImage( mainImage ) );
+                    }
 
                     if ( AppLovinSdk.VERSION_CODE >= 11_04_00_00 )
                     {