+++ Smaato/src/main/java/com/applovin/mediation/adapters/SmaatoMediationAdapter.java
@@ -610,6 +610,7 @@ public class SmaatoMediationAdapter
                     }
 
                     ImageView maxNativeAdMediaView = null;
+                    MaxNativeAd.MaxNativeAdImage maxNativeMainImage = null;
                     if ( assets.images().size() > 0 )
                     {
                         NativeAdAssets.Image image = assets.images().get( 0 );
@@ -617,6 +618,7 @@ public class SmaatoMediationAdapter
                         {
                             maxNativeAdMediaView = new ImageView( context );
                             maxNativeAdMediaView.setImageDrawable( image.drawable() );
+                            maxNativeMainImage = new MaxNativeAd.MaxNativeAdImage( image.drawable() );
                         }
                     }
 
@@ -628,6 +630,10 @@ public class SmaatoMediationAdapter
                             .setBody( assets.text() )
                             .setMediaView( maxNativeAdMediaView )
                             .setCallToAction( assets.cta() );
+                    if ( AppLovinSdk.VERSION_CODE >= 11_04_03_99 )
+                    {
+                        builder.setMainImage( maxNativeMainImage );
+                    }
                     MaxNativeAd maxNativeAd = new MaxSmaatoNativeAd( builder );
 
                     log( "Native ad fully loaded: " + placementId );