+++ MyTarget/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 5.15.1.6
+* Add support for returning the main image asset in `MaxNativeAd` for native ads.
+
 ## 5.15.1.5
 * Add support for providing native media content aspect ratio in `MaxNativeAdView`.